requirements.txt
fastapi>=0.100.0
uvicorn>=0.27.0
supabase>=2.0.0
python-multipart
boto3
pydantic>=2.0.0
python-dotenv
gunicorn
requests
aiosmtpd
